+++ src/app/dialogs/new-profile-dialog/new-profile-dialog.component.ts
@@ -5,7 +5,6 @@ import {
   Validators,
   FormGroupDirective
 } from '@angular/forms';
-import { ProfileService } from '../../profile/services/profile.service';
 import { Observable } from 'rxjs';
 import { DialogService } from '../services/dialog.service';
 
@@ -20,11 +19,7 @@ export class NewProfileDialogComponent implements OnInit {
 
   profiles: Observable<string[]>;
 
-  constructor(
-    private fb: FormBuilder,
-    private profileService: ProfileService,
-    private dialog: DialogService
-  ) {}
+  constructor(private fb: FormBuilder, private dialog: DialogService) {}
 
   ngOnInit() {
     this.form = this.fb.group({
@@ -32,9 +27,11 @@ export class NewProfileDialogComponent implements OnInit {
       copyFrom: ''
     });
 
-    this.profiles = this.profileService.profileNames$;
-
     this.dialog.dialogReady();
+
+    this.dialog.dialogInput.subscribe(profiles => {
+      this.profiles = profiles;
+    });
   }
 
   createProfile() {