+++ src/app/main/package-selection/package-table/package-table.component.spec.ts
@@ -23,6 +23,7 @@ import { PreferencesService } from '../../../core/services/preferences.service';
 import { HumanizePipe } from '../../../shared/humanize.pipe';
 import { PackageService } from '../../services/package.service';
 import { RouterTestingModule } from '@angular/router/testing';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-test-host',
@@ -37,6 +38,14 @@ class TestHostComponent {
 })
 class MockFocusOnKeysDirective {
   @Input('appFocusOnKeys') key: any;
+}
+
+class TranslateServiceStub {
+  setDefaultLang(lang: string): void {}
+  use(lang: string): void {}
+  get() {
+    return of();
+  }
 }
 
 describe('PackageTableComponent', () => {
@@ -66,12 +75,18 @@ describe('PackageTableComponent', () => {
         { provide: PreferencesService, useClass: MockPreferencesService },
         { provide: ElectronService, useClass: MockElectronService },
         { provide: HttpClient, useClass: MockHttpClient },
-        { provide: PackageService, useClass: MockPackageService }
+        { provide: PackageService, useClass: MockPackageService },
+        { provide: TranslateService, useClass: TranslateServiceStub }
       ]
     }).compileComponents();
 
     const http = TestBed.get(HttpClient);
     spyOn(http, 'get').and.returnValue(of([]));
+
+    const prefs: PreferencesService = TestBed.get(PreferencesService);
+    spyOn(prefs, 'get').and.callFake(key =>
+      key === 'displayedColumns' ? [] : null
+    );
   }));
 
   beforeEach(() => {