+++ src/app/main/package-selection/packages-page/step-two/step-two.component.ts
@@ -43,11 +43,15 @@ export class StepTwoComponent implements OnInit, OnDestroy {
     this.subscription.add(
       this.formStep2
         .get('confirmed')
-        .valueChanges.subscribe(this.confirmed.emit.bind(this.confirmed))
+        .valueChanges.subscribe(this.onConfirmChange)
     );
   }
 
   ngOnDestroy() {
     this.subscription.unsubscribe();
   }
+
+  onConfirmChange = (confirmed: boolean) => {
+    this.confirmed.emit();
+  };
 }